+++ examples/controllers/admin/SettingsController.ts
@@ -1,8 +1,10 @@
 import express from 'express';
-import { Controller, Path, Get, Put } from '../../../src';
+import {Controller, Path, Get, Put, FailRoute, Name} from '../../../src';
+import TerminalSettingsApiController from "./settings/TerminalSettingsApiController";
 
 @Path('/settings')
-class AdminSettingsController extends Controller {
+@Name('my-settings')
+export default class AdminSettingsController extends Controller {
   middlewareAdminOnly(
     req: express.Request,
     _res: express.Response,
@@ -26,6 +28,13 @@ class AdminSettingsController extends Controller {
   updateSettings() {
     return { updated: true };
   }
-}
 
-export default AdminSettingsController;
+  @FailRoute
+  get404() {
+    return { error: 'Not Found' };
+  }
+
+  groupTerminal() {
+    return TerminalSettingsApiController;
+  }
+}