+++ examples/controllers/admin/settings/TerminalSettingsApiController.ts
@@ -1,0 +1,10 @@
+import {Controller, Path} from "../../../../src";
+
+@Path('/terminal')
+export default class TerminalSettingsApiController extends Controller {
+  get() {
+    return {
+      status: 'ok',
+    };
+  }
+}